Split conftest to allow reusing the code in other packages
